+++ api/src/error.rs
@@ -1,19 +1,13 @@
 use crate::parser::Error as ParserError;
 
-#[derive(Clone, Debug, Fail, PartialEq)]
+#[derive(Clone, Debug, thiserror::Error, PartialEq)]
 pub enum Error {
-    #[fail(display = "Units are not compatible: {}, {}", lhs, rhs)]
+    #[error("Units are not compatible: {lhs:?}, {rhs:?}")]
     IncompatibleUnitTypes { lhs: String, rhs: String },
 
-    #[fail(display = "{}", _0)]
-    ParsingFailed(#[fail(cause)] ParserError),
+    #[error("{0}")]
+    ParsingFailed(#[from] ParserError),
 
-    #[fail(display = "Operation caused a divide by 0")]
+    #[error("Operation caused a divide by 0")]
     DivideByZero,
-}
-
-impl From<ParserError> for Error {
-    fn from(other: ParserError) -> Self {
-        Self::ParsingFailed(other)
-    }
 }